Updates to A Star
- Updated main so that our test enemy will have their path updated every 3 seconds to go towards the mouse position. If the mouse isn't in a valid position at the time, it will go to the last valid position instead.

- Updated AStar test object movement to be based on position rather than tilemap. Object should move closer to the center of the tile and have an easier time going the right way when given a new path.

- Fixed a bug with AStar that caused two tiles to list each other as their parents, resulting in a path with an infinite length and freezing the program.

+++ src/tilemap/tilemap.hpp
@@ -20,9 +20,9 @@ void setTilemap(const char *filename) ;
 
 void getTilemap(const char *filename) ;
 
-int getTilePos(int pos) ;
+int getTilePos(float pos) ;
 
-bool isWall(int xpos, int ypos) ;
+bool isWall(float xpos, float ypos) ;
 
 int getRows() ;
